feat: 프로젝트 페이지 스켈레톤 추가
